Splitted provider and client for brevity
